services/meta-notifications-client/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.12.5
pydantic_core==2.41.5
